+++ examples/yew-shadow/src/main.rs
@@ -1,6 +1,6 @@
 use stylist::manager::StyleManager;
 use stylist::yew::Global;
-use stylist::{IntoStyle, Style, YieldStyle};
+use stylist::{Style, StyleSource, YieldStyle};
 use web_sys::{window, Element, ShadowRootInit, ShadowRootMode};
 use yew::prelude::*;
 
@@ -133,7 +133,7 @@ impl Component for App {
 }
 
 impl YieldStyle for App {
-    fn style_from(&self) -> IntoStyle {
+    fn style_from(&self) -> StyleSource<'static> {
         r#"
             box-shadow: 0 0 5px 1px rgba(0, 0, 0, 0.7);
             height: 500px;