+++ packages/stylist-macros/src/css.rs
@@ -4,5 +4,5 @@ use quote::quote;
 pub(crate) fn macro_fn(input: TokenStream) -> TokenStream {
     let sheet_tokens = crate::sheet::macro_fn(input);
 
-    quote! { ::stylist::IntoStyle::Sheet(#sheet_tokens) }
+    quote! { ::stylist::StyleSource::from(#sheet_tokens) }
 }