+++ packages/stylist/src/ast/mod.rs
@@ -32,8 +32,6 @@ pub use stylist_core::ast::*;
 #[cfg(feature = "macros")]
 pub use stylist_macros::sheet;
 
-mod into_sheet;
 mod sheet_ref;
 
-pub use into_sheet::IntoSheet;
 pub use sheet_ref::SheetRef;