+++ packages/stylist/src/global_style.rs
@@ -1,12 +1,12 @@
 use std::rc::Rc;
 
-use crate::ast::{IntoSheet, SheetRef, ToStyleStr};
+use crate::ast::{SheetRef, ToStyleStr};
 use crate::manager::StyleManager;
 use crate::registry::StyleKey;
 use crate::style::StyleContent;
 use crate::style::StyleId;
 use crate::utils::get_entropy;
-use crate::Result;
+use crate::{Result, StyleSource};
 
 /// A struct that represents a global Style.
 ///
@@ -22,8 +22,10 @@ pub struct GlobalStyle {
 impl GlobalStyle {
     // The big method is monomorphic, so less code duplication and code bloat through generics
     // and inlining
-    fn create_impl(css: SheetRef, manager: StyleManager) -> Result<Self> {
+    fn create_impl(css: StyleSource<'_>, manager: StyleManager) -> Result<Self> {
         let prefix = format!("{}-global", manager.prefix());
+
+        let css = css.try_to_sheet()?;
 
         // Creates the StyleKey, return from registry if already cached.
         let key = StyleKey {
@@ -75,22 +77,21 @@ impl GlobalStyle {
     /// let style = Style::new("background-color: red;")?;
     /// # Ok::<(), stylist::Error>(())
     /// ```
-    pub fn new<Css>(css: Css) -> Result<Self>
+    pub fn new<'a, Css>(css: Css) -> Result<Self>
     where
-        Css: IntoSheet,
+        Css: Into<StyleSource<'a>>,
     {
         Self::new_with_manager(css, StyleManager::default())
     }
 
     /// Creates a new style using a custom manager.
-    pub fn new_with_manager<Css, M>(css: Css, manager: M) -> Result<Self>
+    pub fn new_with_manager<'a, Css, M>(css: Css, manager: M) -> Result<Self>
     where
-        Css: IntoSheet,
+        Css: Into<StyleSource<'a>>,
         M: Into<StyleManager>,
     {
-        let css = css.into_sheet()?;
         let mgr = manager.into();
-        Self::create_impl(css, mgr)
+        Self::create_impl(css.into(), mgr)
     }
 
     /// Get the parsed and generated style in `&str`.